+++ DesignerMacros/src/common/crudHelper.ts
@@ -8,6 +8,185 @@ class CrudConstants {
 
 class CrudHelper {
 
+    // Super basic selection dialog.
+    public static async openBasicSelectEntityDialog(options?: IISelectEntityDialogOptions): Promise<MacroApi.Context.IElementApi> {
+        let classes = lookupTypesOf("Class").filter(x => CrudHelper.filterClassSelection(x, options));
+        if (classes.length == 0) {
+            await dialogService.info("No Domain types could be found. Please ensure that you have a reference to the Domain package and that at least one class exists in it.");
+            return null;
+        }
+
+        let classId = await dialogService.lookupFromOptions(classes.map((x) => ({
+            id: x.id,
+            name: getFriendlyDisplayNameForClassSelection(x),
+            additionalInfo: `(${x.getParents().map(item => item.getName()).join("/")})`
+        })));
+
+        if (classId == null) {
+            await dialogService.error(`No class found with id "${classId}".`);
+            return null;
+        }
+
+        let foundEntity = lookup(classId);
+        return foundEntity;
+
+        function getFriendlyDisplayNameForClassSelection(element: MacroApi.Context.IElementApi): string {
+            let aggregateEntity = DomainHelper.getOwningAggregate(element);
+            return !aggregateEntity ? element.getName() : `${element.getName()} (${aggregateEntity.getName()})`;
+        }
+    }
+
+    public static async openCrudCreationDialog(options?: IISelectEntityDialogOptions): Promise<ICrudCreationResult|null> {
+        let classes = lookupTypesOf("Class").filter(x => CrudHelper.filterClassSelection(x, options));
+        if (classes.length == 0) {
+            await dialogService.info("No Domain types could be found. Please ensure that you have a reference to the Domain package and that at least one class exists in it.");
+            return null;
+        }
+        
+        let dialogResult = await dialogService.openForm({
+            title: "CRUD Creation Options",
+            fields: [
+                {
+                    id: "entityId",
+                    fieldType: "select",
+                    label: "Entity for CRUD operations",
+                    selectOptions: classes.map(x => {
+                        return {
+                            id: x.id,
+                            description: x.getName(),
+                            additionalInfo: getClassAdditionalInfo(x)
+                        };
+                    }),
+                    isRequired: true
+                },
+                {
+                    id: "create",
+                    fieldType: "checkbox",
+                    label: "Create",
+                    value: "true",
+                    hint: "Generate the \"Create\" operation"
+                },
+                {
+                    id: "update",
+                    fieldType: "checkbox",
+                    label: "Update",
+                    value: "true",
+                    hint: "Generate the \"Update\" operation"
+                },
+                {
+                    id: "queryById",
+                    fieldType: "checkbox",
+                    label: "Query By Id",
+                    value: "true",
+                    hint: "Generate the \"Query By Id\" operation"
+                },
+                {
+                    id: "queryAll",
+                    fieldType: "checkbox",
+                    label: "Query All",
+                    value: "true",
+                    hint: "Generate the \"Query All\" operation"
+                },
+                {
+                    id: "delete",
+                    fieldType: "checkbox",
+                    label: "Delete",
+                    value: "true",
+                    hint: "Generate the \"Delete\" operation"
+                },
+                {
+                    id: "domain",
+                    fieldType: "checkbox",
+                    label: "Domain Operations",
+                    value: "true",
+                    hint: "Generate operations for Domain Entity operations"
+                }
+            ]
+        });
+
+        let foundEntity = lookup(dialogResult.entityId);
+
+        var result: ICrudCreationResult = {
+            selectedEntity: foundEntity,
+            canCreate: dialogResult.create == "true",
+            canUpdate: dialogResult.update == "true",
+            canQueryById: dialogResult.queryById == "true",
+            canQueryAll: dialogResult.queryAll == "true",
+            canDelete: dialogResult.delete == "true",
+            canDomain: dialogResult.domain == "true",
+            selectedDomainOperationIds: []
+        };
+
+        if (result.canDomain && foundEntity.getChildren("Operation").length > 0) {
+            dialogResult = await dialogService.openForm({
+                title: "Select Domain Operations",
+                fields: [
+                    {
+                        id: "tree",
+                        fieldType: "tree-view",
+                        label: "Domain Operations",
+                        hint: "Generate operations from selected domain entity operations",
+                        treeViewOptions: {
+                            rootId: foundEntity.id,
+                            submitFormTriggers: ["double-click", "enter"],
+                            isMultiSelect: true,
+                            selectableTypes: [
+                                {
+                                    specializationId: "Class",
+                                    autoExpand: true,
+                                    autoSelectChildren: false,
+                                    isSelectable: (x) => false
+                                },
+                                {
+                                    specializationId: "Operation",
+                                    isSelectable: (x) => true
+                                }
+                            ]
+                        }
+                    }
+                ]
+            });
+
+            result.selectedDomainOperationIds = dialogResult.tree?.filter((x:any) => x != "0") ?? [];
+        }
+
+        return result;
+
+        function getClassAdditionalInfo(element: MacroApi.Context.IElementApi): any {
+            let aggregateEntity = DomainHelper.getOwningAggregate(element);
+            let prefix = aggregateEntity ? `: ${aggregateEntity.getName()}  ` : "";
+            return `${prefix}(${element.getParents().map(item => item.getName()).join("/")})`;
+        }
+    }
+
+    public static filterClassSelection(element: MacroApi.Context.IElementApi, options?: IISelectEntityDialogOptions) : boolean {
+
+        if (!(options?.allowAbstract ?? false) && element.getIsAbstract()){
+            return false;
+        }
+
+        if (element.hasStereotype("Repository")){
+            return true;
+        }
+
+        if (options?.includeOwnedRelationships != false && DomainHelper.ownerIsAggregateRoot(element)){
+            return  DomainHelper.hasPrimaryKey(element);
+        }
+
+
+        if (DomainHelper.isAggregateRoot(element)){
+            let generalizations = element.getAssociations("Generalization").filter(x => x.isTargetEnd());
+            if (generalizations.length == 0) {
+                return true;
+            }
+            let generalization = generalizations[0];
+            let parentEntity = generalization.typeReference.getType();
+            //Could propagate options here but then we need to update compositional crud to support inheritance and it's already a bit of a hack
+            return CrudHelper.filterClassSelection(parentEntity, {includeOwnedRelationships: false, allowAbstract: true});
+        }
+        return false;
+    } 
+
     static getName(command: MacroApi.Context.IElementApi, mappedElement: MacroApi.Context.IElementApi, dtoPrefix: string = null): string{
         if (mappedElement.typeReference != null) 
             mappedElement = mappedElement.typeReference.getType();
@@ -280,8 +459,23 @@ class CrudHelper {
     // to domain entities that are within a Cosmos DB paradigm.
     static legacyPartitionKey(attribute: MacroApi.Context.IElementApi): boolean {
         return attribute.hasStereotype("Partition Key") && attribute.getName() === "PartitionKey";
-    }
-    
+    }   
+}
+
+interface IISelectEntityDialogOptions {
+    includeOwnedRelationships: boolean;
+    allowAbstract?: boolean;
+}
+
+interface ICrudCreationResult {
+    selectedEntity: MacroApi.Context.IElementApi,
+    selectedDomainOperationIds: string[],
+    canCreate: boolean,
+    canUpdate: boolean,
+    canQueryById: boolean,
+    canQueryAll: boolean,
+    canDelete: boolean,
+    canDomain: boolean
 }
 
 interface IAttributeWithMapPath {