+++ DesignerMacros/src/services-crud/create-crud-advanced-mapping/create-crud-macro-advanced-mapping.ts
@@ -3,12 +3,19 @@
 
 async function execute(element: IElementApi) {
     const package = element.getPackage();
-    const entity = await DomainHelper.openSelectEntityDialog({
-        includeOwnedRelationships: false
-    });
-    if (!entity) { return; }
 
-    if (privateSettersOnly && !hasConstructor(entity)) {
+    let dialogResult = await CrudHelper.openCrudCreationDialog({
+        includeOwnedRelationships: false, // App Service doesn't support nested (yet)
+        allowAbstract: false
+    });
+
+    if (!dialogResult) {
+        return;
+    }
+
+    let entity: MacroApi.Context.IElementApi = dialogResult.selectedEntity;
+
+    if (privateSettersOnly && !hasConstructor(entity) && dialogResult.canCreate) {
         await dialogService.warn(
 `Partial Service Operation Creation.
 Some service operations were created successfully, but was limited due to private setters being enabled, and no constructor is present for entity '${entity.getName()}'.
@@ -17,42 +24,66 @@ To avoid this limitation in the future, either disable private setters or add a 
     }
 
     const serviceName = `${toPascalCase(pluralize(DomainHelper.ownerIsAggregateRoot(entity) ? DomainHelper.getOwningAggregate(entity).getName() : entity.getName()))}Service`;
-    const existingService = element.specialization == "Service" ? element : package.getChildren("Service").find(x => x.getName() == pluralize(serviceName));
+    const existingService = element.specialization == "Service" ? element : package.getChildren("Service").find(x => x.getName() == serviceName);
     const service = existingService ?? createElement("Service", serviceName, package.id);
 
-    const folderName = pluralize(DomainHelper.ownerIsAggregateRoot(entity) ? DomainHelper.getOwningAggregate(entity).getName() : entity.getName());
-    const existingFolder = package.getChildren("Folder").find(x => x.getName() == pluralize(folderName));
-    const folder = existingFolder ?? createElement("Folder", pluralize(folderName), package.id);
+    const targetFolder = getOrCreateEntityFolder(element, entity);
     const primaryKeys = DomainHelper.getPrimaryKeys(entity);
 
-    const resultDto = servicesCrud.createMappedResultDto(entity, folder);
-
-    if (!privateSettersOnly || hasConstructor(entity)) {
-        servicesCrud.createStandardCreateOperation(service, entity, folder);
+    let resultDto: MacroApi.Context.IElementApi = null;
+    if (dialogResult.canQueryById || dialogResult.canQueryAll) {
+        resultDto = servicesCrud.createMappedResultDto(entity, targetFolder);
     }
 
-    if (primaryKeys.length > 0) {
+    if ((!privateSettersOnly || hasConstructor(entity)) && dialogResult.canCreate) {
+        servicesCrud.createStandardCreateOperation(service, entity, targetFolder);
+    }
+
+    if (primaryKeys.length > 0 && (!privateSettersOnly && dialogResult.canUpdate)) {
+        servicesCrud.createStandardUpdateOperation(service, entity, targetFolder);
+    }
+
+    if (primaryKeys.length > 0 && dialogResult.canQueryById) {
         servicesCrud.createStandardFindByIdOperation(service, entity, resultDto);
     }
 
-    servicesCrud.createStandardFindAllOperation(service, entity, resultDto);
+    if (dialogResult.canQueryAll) {
+        servicesCrud.createStandardFindAllOperation(service, entity, resultDto);
+    }
 
     if (primaryKeys.length > 0) {
-        if (!privateSettersOnly){
-            servicesCrud.createStandardUpdateOperation(service, entity, folder);
-        }
-        servicesCrud.createStandardDeleteOperation(service, entity);
-
-        const operations = DomainHelper.getCommandOperations(entity);     
-        for (const operation of operations) {
-            servicesCrud.createCallOperationCommand(service, operation, entity, folder);
+        if (dialogResult.canDelete) {
+            servicesCrud.createStandardDeleteOperation(service, entity);
         }
 
+        if (dialogResult.canDomain) {
+            const operations = DomainHelper.getCommandOperations(entity);     
+            for (const operation of operations) {
+                if (!dialogResult.selectedDomainOperationIds.some(x => x == operation.id)) {
+                    continue;
+                }
+                servicesCrud.createCallOperationCommand(service, operation, entity, targetFolder);
+            }
+        }
     }
 
     service.getChildren("Operation").forEach(operation => {
         convertToAdvancedMapping.convertOperation(operation, entity);
     })
+}
+
+function getAggregateRootFolderName(entity: MacroApi.Context.IElementApi) {
+    return pluralize(DomainHelper.ownerIsAggregateRoot(entity) ? DomainHelper.getOwningAggregate(entity).getName() : entity.getName());
+}
+
+function getOrCreateEntityFolder(folderOrPackage: MacroApi.Context.IElementApi, entity: MacroApi.Context.IElementApi): MacroApi.Context.IElementApi {
+    if (folderOrPackage.specialization == "Folder") {
+        return element;
+    }
+
+    const folderName = getAggregateRootFolderName(entity);
+    const folder = element.getChildren().find(x => x.getName() == pluralize(folderName)) ?? createElement("Folder", pluralize(folderName), element.id);
+    return folder;
 }
 
 function hasConstructor(entity: MacroApi.Context.IElementApi): boolean {