+++ Modules/Intent.Modules.Common.CSharp/release-notes.md
@@ -1,6 +1,7 @@
 ﻿### Version 3.8.6
 
 - Fixed: Mappings not discovering and resolving their C# references in certain circumstances.
+- Improvement: Changed the `.ConcatCode(...)` extension method from `internal` to `public`.
 
 ### Version 3.8.5